eststst

+++ webhard/src/main/MainPage.java
@@ -70,16 +70,16 @@ import javax.swing.border.MatteBorder;
 import javax.swing.border.SoftBevelBorder;
 
 public class MainPage extends JFrame {
-	
+
 	private AccessList Al;
 	private UserList ul;
-	
+
 	private CompanyList CompList;
 	private CompanyInsert ci;
-	
+
 	private FolderInsert folderInsert;
 	private FolderUpdate folderupdate;
-	
+
 	private Button Logout;
 	private JTable table;
 	private JPanel panel_2;
@@ -116,7 +116,7 @@ public class MainPage extends JFrame {
 	 * 
 	 * @param access
 	 */
-	public MainPage(String name,String companyname, final String id, int admin, int access) {
+	public MainPage(String name,final String companyname, final String id, int admin, int access) {
 		if(access==0){
 			JOptionPane.showMessageDialog(null,id+"님은 안증대기 상태입니다.");
 		}
@@ -142,9 +142,14 @@ public class MainPage extends JFrame {
 				
 				public void actionPerformed(ActionEvent e) {
 						
+					CompanyDao ComDao = new CompanyDao();
+					FolderDao folDao = new FolderDao();
+					int UserCompNum = ComDao.selectCompanyNum(companyname);
+					int compNum = folDao.selectcompanyNum(parentNum);
 					
 					if(folderInsert == null){	
-						
+						if(parentNum != 140){
+							if(UserCompNum == compNum || id.equals("admin")){
 						if (parentNum != 0) {
 							selectNode = (DefaultMutableTreeNode) tree.getLastSelectedPathComponent();
 							folderInsert = new FolderInsert(parentNum,id, companyNum, MainPage.this);
@@ -156,6 +161,12 @@ public class MainPage extends JFrame {
 							
 						} else {
 							JOptionPane.showMessageDialog(null, "폴더를 선택해주세요.");
+						}
+							}else{
+								JOptionPane.showMessageDialog(null, "등록할수없는 회사입니다.");
+							}
+						}else{
+							JOptionPane.showMessageDialog(null, "Home에서는 폴더 생성을 할수없습니다.");
 						}
 					}else{
 						JOptionPane.showMessageDialog(null,"이미 사용중인 서비스입니다.");
@@ -171,9 +182,16 @@ public class MainPage extends JFrame {
 				@Override
 				public void actionPerformed(ActionEvent e) {
 					
+					CompanyDao ComDao = new CompanyDao();
+					FolderDao folDao = new FolderDao();
+					int UserCompNum = ComDao.selectCompanyNum(companyname);
+					int compNum = folDao.selectcompanyNum(parentNum);
+					int selectParent = folDao.parentHoemNum(parentNum);
 					if(folderupdate == null){
-						
 						if (parentNum != 0) {
+						if(selectParent != 140){
+						if(UserCompNum == compNum || id.equals("admin")){
+						if(parentNum != 140){
 							folderupdate = new FolderUpdate(parentNum,MainPage.this, companyNum, id);
 							Dimension dim = Toolkit.getDefaultToolkit().getScreenSize();
 							folderupdate.setLocation((dim.width / 2) - (folderupdate.getWidth() / 2),
@@ -181,30 +199,57 @@ public class MainPage extends JFrame {
 							folderupdate.setVisible(true);
 								
 						} else {
+							JOptionPane.showMessageDialog(null, "수정할수없는 폴더입니다.");
+						}
+						}else{
+							JOptionPane.showMessageDialog(null, "수정할수없는 폴더입니다.");
+						}
+						}else{
+							JOptionPane.showMessageDialog(null, "수정할수없는 폴더입니다.");
+						}
+						}else{
 							JOptionPane.showMessageDialog(null, "폴더를 선택해주세요.");
 						}
 					}else{
 						JOptionPane.showMessageDialog(null,"이미 사용중인 서비스입니다.");
 					}
-				}
+				}			
 			});
 			forder.add(update);
 
 			delete = new JMenuItem("폴더 삭제");
 			delete.addActionListener(new ActionListener() {
-
+				
 				@Override
 				public void actionPerformed(ActionEvent e) {
+					
 					FolderDao dao = new FolderDao();
-
-					if (parentNum != 0) {
-						selectNode = (DefaultMutableTreeNode) tree
-								.getLastSelectedPathComponent();
-						deleteFolder(parentNum);
-						tree.updateUI();
-					} else {
-						JOptionPane.showMessageDialog(null, "폴더를 선택해주세요.");
+					CompanyDao ComDao = new CompanyDao();
+					int UserCompNum = ComDao.selectCompanyNum(companyname);
+					int compNum = dao.selectcompanyNum(parentNum);
+					int selectParent = dao.parentHoemNum(parentNum);
+					System.out.println(selectParent);
+					System.out.println(parentNum);
+					if(parentNum == 140){
+						JOptionPane.showMessageDialog(null, "Home폴더는 삭제하실 수 없습니다.");
+					}else{
+						if (parentNum != 0) {
+							if(selectParent != 140 ){
+								if(UserCompNum == compNum || id.equals("admin")){
+							selectNode = (DefaultMutableTreeNode) tree.getLastSelectedPathComponent();
+							deleteFolder(parentNum);
+							tree.updateUI();
+							}else{
+								JOptionPane.showMessageDialog(null, "삭제할수없는 폴더업니다.");
+							}
+						} else {
+							JOptionPane.showMessageDialog(null, "삭제할수없는 폴더업니다.");
+						}	
+						}else{
+							JOptionPane.showMessageDialog(null, "폴더를 선택해주세요.");
+						}
 					}
+					
 
 				}
 			});
@@ -216,6 +261,7 @@ public class MainPage extends JFrame {
 			JMenuItem flieinsert = new JMenuItem("파일 등록");
 			flieinsert.addActionListener(new ActionListener() {
 				public void actionPerformed(ActionEvent e) {
+					if(parentNum != 140){	
 					if (parentNum != 0) {
 						selectNode = (DefaultMutableTreeNode) tree.getLastSelectedPathComponent();
 						FileInsert flieI = new FileInsert(parentNum,companyNum, id, MainPage.this);
@@ -225,6 +271,9 @@ public class MainPage extends JFrame {
 						flieI.setVisible(true);
 					} else {
 						JOptionPane.showMessageDialog(null, "폴더를 선택해주세요.");
+					}
+					}else{
+						JOptionPane.showMessageDialog(null, "Home폴더에는 추가 하실수없습니다.");
 					}
 				}
 			});
@@ -236,7 +285,7 @@ public class MainPage extends JFrame {
 
 				@Override
 				public void actionPerformed(ActionEvent e) {
-
+					
 					if (parentNum != 0) {
 						deleteFolder(parentNum);
 						tree.updateUI();
@@ -501,6 +550,7 @@ public class MainPage extends JFrame {
 						int parentNum1 = selectObtion.getCompanyNum();
 						// int Home = selectObtion.getItemNum();
 						parentNum = selectObtion.getItemNum();
+					
 						companyNum = selectObtion.getCompanyNum();
 
 						int selectpnum = cdao.selectCompanyNum(compname);
@@ -524,14 +574,28 @@ public class MainPage extends JFrame {
 										.printFileInParentFolder(parentNum);
 
 								listBySearch(folders, fileList);
-
+								
+								boolean filecheck = fileDao.checkfile(parentNum);
+								boolean foldercheck = folderDao.checkFolder(parentNum);
 								// //////////////////////////////////
 								if (parentNode != null)
 									pobj = parentNode.getUserObject();
-								if (((e.getModifiers() & InputEvent.BUTTON3_MASK) != 0)
-										&& (tree.getSelectionCount() > 0)) {
-									showMenu(e.getX(), e.getY(), id);
+								if(parentNum != 140){
+								if(filecheck == true && foldercheck == false){
+									if (((e.getModifiers() & InputEvent.BUTTON3_MASK) != 0)
+											&& (tree.getSelectionCount() > 0)) {
+										showfileMenu(e.getX(), e.getY(), id);
+									}
+								}else if(filecheck == false && foldercheck == true){
+									if (((e.getModifiers() & InputEvent.BUTTON3_MASK) != 0)
+											&& (tree.getSelectionCount() > 0)) {
+										showfolderMenu(e.getX(), e.getY(), id);
+									}
 								}
+								}else{
+									JOptionPane.showMessageDialog(null,"Home폴더는 사용하실수 없습니다.");
+								}
+								
 							}
 						}
 					}
@@ -593,8 +657,7 @@ public class MainPage extends JFrame {
 	}
 
 	public void deleteFolder(int itemNum) {
-		
-		
+
 		ArrayList<Integer> childs = new ArrayList<Integer>();
 		FolderDao dao = new FolderDao();
 		childs = (ArrayList<Integer>) dao.itemNumByParentNum(itemNum);
@@ -619,7 +682,7 @@ public class MainPage extends JFrame {
 			dao.deleteFolder(parentNum);
 
 		}
-		
+
 		selectNode.removeAllChildren();
 		selectNode.removeFromParent();
 	}
@@ -673,7 +736,7 @@ public class MainPage extends JFrame {
 		table.updateUI();
 	}
 
-	protected void showMenu(int x, int y, final String id) {
+	protected void showfolderMenu(int x, int y, final String id) {
 		JPopupMenu popup = new JPopupMenu();
 		JMenuItem mi = new JMenuItem("폴더 추가");
 		popup.add(mi);
@@ -704,8 +767,13 @@ public class MainPage extends JFrame {
 				deleteSelectedItems();
 			}
 		});
+		popup.show(tree, x, y);
+	}
 
-		mi = new JMenuItem("파일 다운로드");
+	protected void showfileMenu(int x, int y, final String id) {
+
+		JPopupMenu popup = new JPopupMenu();
+		JMenuItem mi = new JMenuItem("파일 다운로드");
 		popup.add(mi);
 		mi.addActionListener(new ActionListener() {
 
@@ -726,7 +794,6 @@ public class MainPage extends JFrame {
 				} else {
 					JOptionPane.showMessageDialog(null, "폴더를 선택해주세요.");
 				}
-
 			}
 		});
 		popup.show(tree, x, y);
@@ -824,6 +891,7 @@ public class MainPage extends JFrame {
 		tree.updateUI();
 
 	}
+
 	public void deleteCompanyFolder(int comNum) {
 		FolderDao dao = new FolderDao();
 		ItemDto itemdto = dao.printFolderbyCompanyNum(comNum);
@@ -851,50 +919,50 @@ public class MainPage extends JFrame {
 		selectNode.add(childNode);
 		tree.updateUI();
 	}
-	
-	//창 한번만 뜨게 만들기
-	public FolderInsert openFolderInsert(FolderInsert values){
-		
+
+	// 창 한번만 뜨게 만들기
+	public FolderInsert openFolderInsert(FolderInsert values) {
+
 		folderInsert = values;
-		
+
 		return folderInsert;
 	}
-	
-	public FolderUpdate openFolderUpdate(FolderUpdate values){
-		
+
+	public FolderUpdate openFolderUpdate(FolderUpdate values) {
+
 		folderupdate = values;
-		
+
 		return folderupdate;
 	}
-	
-	public AccessList openAccessList(AccessList values){
-		
+
+	public AccessList openAccessList(AccessList values) {
+
 		Al = values;
-		
+
 		return Al;
 	}
-	
-	public UserList openUserList(UserList values){
-		
+
+	public UserList openUserList(UserList values) {
+
 		ul = values;
-		
+
 		return ul;
 	}
-	
-	public CompanyInsert openCompanyInsert(CompanyInsert values){
-		
+
+	public CompanyInsert openCompanyInsert(CompanyInsert values) {
+
 		ci = values;
-		
+
 		return ci;
 	}
 
-	public CompanyList openCompanyList(CompanyList values){
-		
+	public CompanyList openCompanyList(CompanyList values) {
+
 		CompList = values;
-		
+
 		return CompList;
 	}
-	
+
 	public class CustomTreeCellRenderer extends DefaultTreeCellRenderer {
 		Icon folderIcon = new ImageIcon("Folder.png");
 		Icon closedIcon = new ImageIcon("Folder.png");
